Move kube controller to the federation package
