+++ pkg/servicemesh/federation/exports/controller.go
@@ -23,8 +23,8 @@ import (
 	"k8s.io/client-go/tools/cache"
 	v1 "maistra.io/api/federation/v1"
 
-	kubecontroller "istio.io/istio/pkg/kube/controller"
 	"istio.io/istio/pkg/servicemesh/federation/common"
+	kubecontroller "istio.io/istio/pkg/servicemesh/federation/kube"
 )
 
 const controllerName = "federation-exports-controller"
